refactor: flatten src/test
